Lists all articles, with working vote controls
